physics update

+++ src/client/viewer.components/Viewer.Extensions.Dynamic/Viewing.Extension.MetaProperties/Viewing.Extension.MetaProperties.js
@@ -3,9 +3,11 @@
 // By Philippe Leefsma, Autodesk Inc, April 2017
 //
 ///////////////////////////////////////////////////////////
+import {ReflexContainer, ReflexElement, ReflexSplitter} from 'react-reflex'
 import MultiModelExtensionBase from 'Viewer.MultiModelExtensionBase'
+import Search from './Viewing.Extension.MetaProperties.Search'
 import MetaAPI from './Viewing.Extension.MetaProperties.API'
-import ExtensionBase from 'Viewer.ExtensionBase'
+import {OverlayTrigger, Popover} from 'react-bootstrap'
 import {AddMetaProperty} from './MetaProperty'
 import WidgetContainer from 'WidgetContainer'
 import MetaTreeView from './MetaTreeView'
@@ -29,18 +31,27 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
 
     this.onPropertyDeleted = this.onPropertyDeleted.bind(this)
     this.onPropertyUpdated = this.onPropertyUpdated.bind(this)
-    this.onPropertyAdded = this.onPropertyAdded.bind(this)
     this.onDeleteProperty = this.onDeleteProperty.bind(this)
+    this.onPropertyAdded = this.onPropertyAdded.bind(this)
     this.onEditProperty = this.onEditProperty.bind(this)
     this.onMetaChanged = this.onMetaChanged.bind(this)
     this.onContextMenu = this.onContextMenu.bind(this)
-    this.renderTitle = this.renderTitle.bind(this)
+    this.onSearch = this.onSearch.bind(this)
 
     this.dialogSvc =
       ServiceManager.getService('DialogSvc')
 
     this.socketSvc =
       ServiceManager.getService('SocketSvc')
+
+    this.react = options.react
+	}
+
+	/////////////////////////////////////////////////////////
+	// Load callback
+  //
+  /////////////////////////////////////////////////////////
+	load () {
 
     this.socketSvc.on (
       'meta.propertyDeleted',
@@ -54,18 +65,10 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
       'meta.propertyAdded',
       this.onPropertyAdded)
 
-    this.react = options.react
-	}
-
-	/////////////////////////////////////////////////////////
-	// Load callback
-  //
-  /////////////////////////////////////////////////////////
-	load () {
-
     this.react.setState({
 
       properties: [],
+      search: false,
       model: null,
       dbId: null
 
@@ -253,6 +256,29 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
   //
   //
   /////////////////////////////////////////////////////////
+  getComponentData (dbId) {
+
+    return new Promise((resolve, reject) => {
+
+      const {model} = this.react.getState()
+
+      model.getProperties(dbId, (result) => {
+
+        resolve ({
+          externalId: result.externalId
+        })
+
+      }, (error) => {
+
+        reject(error)
+      })
+    })
+  }
+
+  /////////////////////////////////////////////////////////
+  //
+  //
+  /////////////////////////////////////////////////////////
   async loadNodeProperties (dbId, refresh) {
 
     const state = this.react.getState()
@@ -277,16 +303,56 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
     const metaProperties =
       await this.api.getNodeMetaProperties(dbId)
 
-    const properties = [
-      ...modelProperties,
-      ...metaProperties
-    ]
+    const properties = this.buildFinalProperties(
+      modelProperties, metaProperties)
 
     await this.react.setState({
       guid: this.guid(),
       properties,
       dbId
     })
+  }
+
+  /////////////////////////////////////////////////////////
+  //
+  //
+  /////////////////////////////////////////////////////////
+  buildFinalProperties (
+    modelProperties, metaProperties) {
+
+    let finalModelProperties = [
+      ...modelProperties
+    ]
+
+    const finalMetaProperties =
+      metaProperties.filter((metaProperty) => {
+
+        if (metaProperty.isOverride) {
+
+          const entries = finalModelProperties.entries()
+
+          for (let [idx, modelProperty] of entries) {
+
+            if (modelProperty.displayCategory ===
+                metaProperty.displayCategory &&
+                modelProperty.displayName ===
+                metaProperty.displayName) {
+
+                finalModelProperties.splice(idx, 1)
+                break
+            }
+          }
+
+          return metaProperty.metaType !== 'DeleteOverride'
+        }
+
+        return true
+      })
+
+    return [
+      ...finalModelProperties,
+      ...finalMetaProperties
+    ]
   }
 
   /////////////////////////////////////////////////////////
@@ -461,38 +527,43 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
   //
   //
   /////////////////////////////////////////////////////////
-  onEditProperty (metaProperty) {
+  onEditProperty (metaProperty, isModelOverride) {
 
     return new Promise ((resolve) => {
 
-      const onClose = (result) => {
+      const onClose = async(result) => {
 
         this.dialogSvc.off('dialog.close', onClose)
 
         if (result === 'OK') {
 
           const newMetaProperty = Object.assign({},
-            this.metaPropertyEdits, {
-              dbId: metaProperty.dbId,
-              id: metaProperty.id
-            })
+            metaProperty, this.metaPropertyEdits)
 
           const newMetaPayload = this.buildMetaPayload(
             newMetaProperty)
 
-          if (newMetaProperty.file && metaProperty.fileId) {
+          if (isModelOverride) {
 
-            this.api.deleteResource(metaProperty.fileId)
+            await this.api.addNodeMetaProperty(
+              newMetaPayload)
+
+          } else {
+
+            if (newMetaProperty.file && metaProperty.fileId) {
+
+              this.api.deleteResource(metaProperty.fileId)
+            }
+
+            await this.api.updateNodeMetaProperty(
+              newMetaPayload)
           }
-
-          this.api.updateNodeMetaProperty(
-            newMetaPayload)
 
           this.socketSvc.broadcast (
             'meta.propertyUpdated',
             newMetaPayload)
 
-          resolve (newMetaPayload)
+          return resolve (newMetaPayload)
         }
 
         resolve (false)
@@ -500,16 +571,21 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
 
       this.dialogSvc.on('dialog.close', onClose)
 
+      this.metaPropertyEdits = {}
+
       const dlgProps = Object.assign({}, metaProperty, {
         disableOK: this.dialogSvc.disableOK,
         onChanged: this.onMetaChanged,
         editMode: true
       })
 
+      const override = isModelOverride
+        ? '[Override] ' : ''
+
       this.dialogSvc.setState({
+        title: `Edit Meta Property ${override}...`,
         content: <AddMetaProperty {...dlgProps}/>,
         className: 'meta-property-dlg',
-        title: 'Edit Meta Property ...',
         disableOK: true,
         open: true
       }, true)
@@ -520,27 +596,41 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
   //
   //
   /////////////////////////////////////////////////////////
-  onDeleteProperty (metaProperty) {
+  onDeleteProperty (metaProperty, isModelOverride) {
 
     return new Promise ((resolve) => {
 
-      const onClose = (result) => {
+      const onClose = async(result) => {
 
         this.dialogSvc.off('dialog.close', onClose)
 
         if (result === 'OK') {
 
-          this.api.deleteNodeMetaProperty(
-            metaProperty.id)
+          if (isModelOverride) {
+
+            const metaPayload = Object.assign({},
+              metaProperty, {
+                metaType: 'DeleteOverride',
+                isOverride: true
+              })
+
+            await this.api.addNodeMetaProperty(
+              metaPayload)
+
+          } else {
+
+            await this.api.deleteNodeMetaProperty(
+              metaProperty.id)
+          }
 
           this.socketSvc.broadcast (
             'meta.propertyDeleted',
             metaProperty.dbId)
 
-          resolve (true)
+          return resolve (metaProperty)
         }
 
-        resolve (false)
+        resolve (null)
       }
 
       this.dialogSvc.on('dialog.close', onClose)
@@ -549,13 +639,75 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
         `Are you sure you want to delete`
         + ` <b>${metaProperty.displayName}</b> ?`)
 
+      const override = isModelOverride
+        ? '[Override] ' : ''
+
       this.dialogSvc.setState({
-        title: 'Delete Property ...',
-        content:
-          <div dangerouslySetInnerHTML={{__html: msg}}>
-          </div>,
+        content: <div dangerouslySetInnerHTML={{__html: msg}}/>,
+        title: `Delete Property ${override}...`,
         open: true
       })
+    })
+  }
+
+  /////////////////////////////////////////////////////////
+  //
+  //
+  /////////////////////////////////////////////////////////
+  deleteAllProperties (dbId) {
+
+    return new Promise ((resolve) => {
+
+      const onClose = async(result) => {
+
+        this.dialogSvc.off('dialog.close', onClose)
+
+        if (result === 'OK') {
+
+          await this.api.deleteNodeMetaProperties(dbId)
+
+          this.loadNodeProperties(dbId, true)
+
+          this.socketSvc.broadcast (
+            'meta.propertyDeleted', dbId)
+
+          return resolve (true)
+        }
+
+        resolve (false)
+      }
+
+      this.dialogSvc.on('dialog.close', onClose)
+
+      const {model} = this.react.getState()
+
+      const instanceTree = model.getData().instanceTree
+
+      const nodeName = instanceTree.getNodeName(dbId)
+
+      const msg = DOMPurify.sanitize(
+        `Are you sure you want to delete all custom properties on`
+        + ` <br/><b>${nodeName}</b> ?`)
+
+      this.dialogSvc.setState({
+        content: <div dangerouslySetInnerHTML={{__html: msg}}/>,
+        title: `Delete Component Properties ...`,
+        open: true
+      })
+    })
+  }
+
+
+  /////////////////////////////////////////////////////////
+  //
+  //
+  /////////////////////////////////////////////////////////
+  onSearch () {
+
+    const {search} = this.react.getState()
+
+    this.react.setState({
+      search: !search
     })
   }
 
@@ -572,6 +724,7 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
       await this.react.popRenderExtension(id)
 
       this.react.pushViewerPanel(this, {
+        className: this.className,
         height: 250,
         width: 350
       })
@@ -590,6 +743,8 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
   /////////////////////////////////////////////////////////
   renderTitle (docked) {
 
+    const {dbId, search} = this.react.getState()
+
     const spanClass = docked
       ? 'fa fa-chain-broken'
       : 'fa fa-chain'
@@ -600,24 +755,38 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
           Meta Properties
         </label>
         <div className="meta-properties-controls">
+          <button className={search ? 'active':''}
+            onClick={this.onSearch}
+            title="Search MetaProperties">
+            <span className="fa fa-search" style={{
+              position: 'relative',
+              top: '-1px'
+            }}/>
+          </button>
+          <OverlayTrigger trigger="click"
+            overlay={this.renderPopover()}
+            placement="left"
+            rootClose>
+            <button title="Export MetaProperties">
+              <span className="fa fa-cloud-download"/>
+            </button>
+          </OverlayTrigger>
+          <button onClick={() => this.showAddMetaPropertyDlg(dbId)}
+            title="Add new MetaProperty on component">
+            <span className="fa fa-plus"/>
+          </button>
+          <button onClick={() => this.deleteAllProperties(dbId)}
+            title="Delete all MetaProperties on component">
+            <span className="fa fa-close" style={{
+              position: 'relative',
+              top: '-1px'
+            }}/>
+          </button>
           <button onClick={() => this.setDocking(docked)}
             title="Toggle docking mode">
             <span className={spanClass}/>
           </button>
         </div>
-      </div>
-    )
-  }
-
-  /////////////////////////////////////////////////////////
-  //
-  //
-  /////////////////////////////////////////////////////////
-  renderControls () {
-
-    return (
-      <div className="controls">
-
       </div>
     )
   }
@@ -636,6 +805,8 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
 
     return (
       <MetaTreeView
+        onDeleteModelProperty={this.onDeleteModelProperty}
+        onEditModelProperty={this.onEditModelProperty}
         menuContainer={this.options.appContainer}
         onDeleteProperty={this.onDeleteProperty}
         onEditProperty={this.onEditProperty}
@@ -652,18 +823,67 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
   //
   //
   /////////////////////////////////////////////////////////
+  renderPopover () {
+
+    const {dbId} = this.react.getState()
+
+    return (
+      <Popover  className={`${this.className} exports`}
+        title="Meta Exports"
+        id="exports-popover" >
+        <button
+          onClick={() => {
+            this.api.exportProperties('json')
+          }}
+          title=".json export">
+          <span className="fa fa-cloud-download"/>
+          <label>
+            MetaProperties.json
+          </label>
+        </button>
+        <button
+          onClick={() => {
+            this.api.exportProperties('csv')
+          }}
+          title=".csv export">
+          <span className="fa fa-cloud-download"/>
+          <label>
+            MetaProperties.csv
+          </label>
+        </button>
+      </Popover>
+    )
+  }
+
+  /////////////////////////////////////////////////////////
+  //
+  //
+  /////////////////////////////////////////////////////////
   renderContent () {
 
-    const {properties} = this.react.getState()
-
-    const content = properties.length
-      ? this.renderTreeView(properties)
-      : <div/>
+    const {properties, search} = this.react.getState()
 
     return (
       <div className="content">
-        <ReactLoader show={!properties.length}/>
-        { content }
+        <ReflexContainer orientation='horizontal'>
+            <ReflexElement>
+              <ReactLoader show={!properties.length}/>
+              {
+                properties.length &&
+                this.renderTreeView(properties)
+              }
+            </ReflexElement>
+            {
+              search &&
+              <ReflexSplitter/>
+            }
+            {
+              search &&
+              <ReflexElement minSize={40}>
+                <Search/>
+              </ReflexElement>
+            }
+          </ReflexContainer>
       </div>
     )
   }
@@ -679,10 +899,7 @@ class MetaPropertiesExtension extends MultiModelExtensionBase {
         renderTitle={() => this.renderTitle(opts.docked)}
         showTitle={opts.showTitle}
         className={this.className}>
-
-        { this.renderControls() }
         { this.renderContent () }
-
       </WidgetContainer>
     )
   }
